# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Instituto Académico — Intranet API
  description: |
    Backend Express + TypeScript + Prisma (MariaDB) para gestión académica.
    ### Autenticación
    - `POST /auth/login` devuelve un **token sin rol**.
    - `POST /auth/select-role` (con ese token) devuelve un **token con rol**.
    - Todas las demás rutas protegen el acceso por rol del token.
    - Envío del token: cabecera `Authorization: Bearer <token>`.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /auth/select-role:
    post:
      tags: [Auth]
      summary: Seleccionar rol de la sesión (devuelve token con rol)
      operationId: selectRole
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SelectRoleRequest' }
      responses:
        '200':
          description: Token con rol seleccionado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelectRoleResponse' }
        '403': { $ref: '#/components/responses/RoleRequired' }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Token JWT con rol (obtenido tras /auth/select-role). Cabecera: Authorization: Bearer <token>"
  responses:
...
    RoleRequired:
      description: El token no tiene rol seleccionado (hacer /auth/select-role)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string, nullable: true, example: Operación exitosa }
        data: { type: object, nullable: true }
      additionalProperties: true
...
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        code: { type: string, nullable: true }
        rolesDisponibles: { type: array, items: { type: string }, nullable: true }

    # -------- Auth --------
...
    SelectRoleRequest:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [ALUMNO, ADMINISTRATIVO, PROFESOR] }
      example:
        role: ADMINISTRATIVO
    SelectRoleResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                token: { type: string, description: Token JWT con rol }
                refreshToken: { type: string }
                sessionId: { type: integer }
                rol: { type: string, enum: [ALUMNO, ADMINISTRATIVO, PROFESOR] }
                rolesDisponibles: { type: array, items: { type: string } }
...